+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.25] - 2025-08-22
+- save task description and notes automatically when editing.
+- still have the bug where the title changes when you edit stuff
+
 ## [0.1.24] - 2025-08-22
 - Persist settings across app restarts.